fix: remove semicolon
This semicolon inhibits links on IRC from being clickable, as it is automatically parsed as part of the link by the webchat.freenode client. By removing it, people can click on the link directly in the chat. I do not think it adds anything substantially, so removing it should be trivial.